% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{iv.est}
\alias{iv.est}
\title{Implied Volatility estimate}
\usage{
iv.est(C, S, t)
}
\arguments{
\item{C}{Call premium}

\item{S}{Spot price of the underlying asset}

\item{t}{Time to expiration in years}
}
\value{
Returns an iv estimate
}
\description{
Estimates the implied volatility of a European-style call option
}
\examples{
iv.est(2.35, 100, (45/365))
}
