% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{putgreek}
\alias{putgreek}
\title{Put Option Greek}
\usage{
putgreek(greek = c("delta", "gamma", "theta", "vega", "rho", "premium"),
  s, x, sigma, t, r, d = 0)
}
\arguments{
\item{greek}{String value, desired option greek to return}

\item{s}{Spot price of the underlying asset}

\item{x}{Strike price of the option}

\item{sigma}{Implied volatility of the underlying asset price, defined as the annualized standard deviation of the asset returns}

\item{t}{Time to maturity in years}

\item{r}{Annual continuously-compounded risk-free rate, use the function r.cont}

\item{d}{Annual continuously-compounded dividend yield, use the function r.cont}
}
\value{
Returns the dired option greek, including premium
}
\description{
Computes the selected option greek, including premium
}
\examples{
putgreek("vega", 100, 100, 0.20, (45/365), 0.02, 0.02)
}
