% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionstrat.R
\name{r.cont}
\alias{r.cont}
\title{Continuously Compounded Rate}
\usage{
r.cont(r, n)
}
\arguments{
\item{r}{nominal rate}

\item{n}{number of times compounded each year}
}
\value{
Returns a continuously compounded rate
}
\description{
Convert a given nominal rate to a continuously compounded rate
}
\examples{
r.cont(0.12, 2)
}
