% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionstrat.R
\name{tdiff}
\alias{tdiff}
\title{Time Difference}
\usage{
tdiff(date1, date2, period = c("days, years"))
}
\arguments{
\item{date1}{Earlier date}

\item{date2}{Later date}

\item{period}{String value, either "days", or "years"}
}
\value{
Returns a numeric value
}
\description{
Computes the difference in time between two dates
}
\examples{
tdiff("2018-01-01", "2018-06-30", "days")
}
